generate crate now builds
