+++ generate/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "ethcontract-generate"
+version = "0.1.0"
+authors = ["Nicholas Rodrigues Lordello <nlordell@gmail.com>"]
+edition = "2018"
+license = "MIT"
+repository = "https://github.com/gnosis/ethcontract-rs"
+homepage = "https://github.com/gnosis/ethcontract-rs"
+documentation = "https://docs.rs/ethcontract"
+description = """
+Code generation for type-safe bindings to Ethereum smart contracts.
+"""
+
+[dependencies]
+anyhow = "1.0"
+ethabi = "8.0"
+ethcontract-common = { version = "0.1.0", path = "../common" }
+Inflector = "0.11"
+proc-macro2 = "1.0"
+quote = "1.0"
+syn = "1.0"
